requirements.txt
pika==1.1.0
pylint==2.3.1
python-dotenv==0.10.3